requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==21.2.0
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0

# Data processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# PDF processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0

# OpenAI/LLM
openai==1.3.5

# Google APIs
google-api-python-client==2.108.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
gspread-dataframe==3.3.1

# Environment and utilities
python-dotenv==1.0.0

# HTTP requests (if needed)
httpx==0.25.1

